Update README.md
Just adding what might be a useful information for anybody interested.

I didn't look further, but I needed to make the repository as small as possible and deleting the .git repository did the job incredibly well. I thought people could use this information. It is very strange that the repo history went this big, it might be interesting to investigate.