Output both crop images and image with boxes
With the argument --both passed to run_detector.py, both the input image with bounding boxes added and the cropped images are output. --crop outputs only cropped images and none of those arguments outputs an image with boxes. It saves time not having to run the script twice for both outputs.