Relative box expansion added
The argument --box_expansion_relative passed to run_detector.py specifies ratio of the number of pixels to the longer side of the boxes to expand them by in addition to the absolute --box_expansion. It is superior for images of animals with different sizes and resolutions having borders with equal proportions. The size of the parts of an animal that be might cut off incorrectly stands in relation to the size of the box as well.