Cropping jpeg images lossless using jpegtran-cffi
If im_file is passed to crop_image and its extension contains 'jp', an autotransformed JPEGImage object is created by jpegtran-cffi.
The box corners are rounded to integers towards the larger box.
The left top corner has to be a multiple of 16 in order to fall on a jpeg block boundary. To achieve an uniform expansion, modulos are added on all sides.
The returned JPEGImage object can be saved be saved the same way as a PIL image.